Move CommonAttentionMetadata
